stm32: add sw workaround for broken stm32 hw RTS implementation
adds support for manually asserting nRTS when the OS buffer is >3/4 full
and deasserting it when it is <1/4 full.

This makes the RTS pin actually work in a useful way (the stm32 hw
implementation is broken - it asserts nRTS as soon as more than one
character is received).

We assert nRTS from inside of up_recvchars.  For deassertion we needed
a new callback from the general serial device layer when characters
have been dequed.  onrxdeque was added to uart_ops_s for this purpose and
implemented (currently) only on the stm32.